Final commit. Adding all the file headers, creating a readme file and removing size_t and changing it to unsigned int

+++ Project-66/console.h
@@ -1,8 +1,22 @@
+/**
+*  
+* Solution to course project # 66
+* Introduction to programming course
+* Faculty of Mathematics and Informatics of Sofia University
+* Winter semester 2025/2026
+*
+* @author Mladen Stankov
+* @idnumber 3MI0600676
+* @compiler GCC
+*
+* <Header file for console commands>
+*
+*/
 #pragma once
 #include "game.h"
 
 
-constexpr size_t MAX_COMMAND_LEN = 100;
+constexpr unsigned int MAX_COMMAND_LEN = 100;
 
 // Returns false if the game is terminated
 void processCommand(char* command, Game& game);
